Added Pinia AuthStore; Minor fixed; Server refactoring;
