+++ client/src/App.vue
@@ -1,10 +1,17 @@
-<script setup>
+<script setup lang="ts">
+import { useAuthStore } from "./stores/auth";
+import NavigationDrawer from "../src/components/NavigationDrawer.vue";
+
+const auth = useAuthStore();
+
 </script>
 
 <template>
-  <v-app>
-    <v-container>
+  <v-app >
+    <NavigationDrawer v-if="auth.isAuthenticated"/>
+    <v-main>
       <router-view />
-    </v-container>
+    </v-main>
   </v-app>
 </template>
+
