+++ client/src/components/ChatMessage.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { defineProps, defineEmits } from "vue";
 
-const props = defineProps<{
+defineProps<{
   message: { _id: string; username: string; content: string };
 }>();
