+++ client/src/components/LikeButton.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { defineProps, defineEmits } from "vue";
 
-const props = defineProps<{
+defineProps<{
   isLiked: boolean;
   likesCount: number;
 }>();
