+++ client/src/components/SubscribeButton.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { defineProps, defineEmits } from "vue";
 
-const props = defineProps<{
+defineProps<{
   isSubscribed: boolean;
 }>();
