+++ client/src/main.ts
@@ -6,6 +6,7 @@ import * as components from "vuetify/components";
 import * as directives from "vuetify/directives";
 import router from "./router";
 import { setRouter } from "./router/routerService";
+import { createPinia } from "pinia";
 
 const vuetify = createVuetify({
   components,
@@ -13,6 +14,7 @@ const vuetify = createVuetify({
 });
 
 const app = createApp(App);
+app.use(createPinia());
 app.use(vuetify);
 app.use(router);
 setRouter(router);
