+++ client/src/pages/AllUsersPage.vue
@@ -4,13 +4,10 @@ import type { IUser, IUsersResponse } from "../models/user.ts";
 import { navigateTo } from "../router/routerService.ts";
 import { Pages } from "../utils/pages.ts";
 import { ref, onMounted, computed } from "vue";
-import NavigationDrawer from "../components/NavigationDrawer.vue";
-import { USER_KEY } from "../utils/constants.ts";
 
 const users = ref<IUser[]>([]);
 const loading = ref<boolean>(true);
 const isEmpty = computed(() => !loading.value && users.value.length === 0);
-const userId = ref<string | null>(localStorage.getItem(USER_KEY));
 
 const vOnGoToDialogue = (user: IUser) => { 
   navigateTo(Pages.Messages, { params: { id: user._id } });
@@ -18,7 +15,7 @@ const vOnGoToDialogue = (user: IUser) => {
 
 const fetchUsers = async (): Promise<void> => {
   try {
-    const response = await axios.get<IUsersResponse>('http://localhost:5004/users');
+    const response = await axios.get<IUsersResponse>('http://localhost:5004/api/users');
     
     if (response.data && Array.isArray(response.data.users)) {
       users.value = response.data.users;
@@ -43,11 +40,7 @@ onMounted(() => {
 </script>
 
 <template>
-  <v-container class="fill-height d-flex">
-    <NavigationDrawer :userId="userId" />
-
-    <v-main class="d-flex" style="min-height: 100vh;">
-      <v-container>
+  <v-container class="fill-height d-flex align-start justify-start">
         <v-row>
           <v-col v-for="user in users" :key="user._id" sm="6" md="4">
             <v-card class="pa-2" style="position: relative" @click="vOnGoToUser(user)">
@@ -69,8 +62,6 @@ onMounted(() => {
         <v-alert class="mt-4" v-if="isEmpty" type="warning">
           Пользователей не найдено.
         </v-alert>
-      </v-container>
-    </v-main>
   </v-container>
 </template>
