+++ server/controllers/auth.controller.js
@@ -1,0 +1,47 @@
+import bcrypt from "bcryptjs";
+import { User } from "../models/User.js";
+
+export const registerUser = async (req, res) => {
+  try {
+    const { username, email, password } = req.body;
+
+    const existingUser = await User.findOne({ username });
+    if (existingUser)
+      return res.status(400).json({ error: "Email is already registered" });
+
+    const salt = await bcrypt.genSalt(10);
+    const hashedPassword = await bcrypt.hash(password, salt);
+
+    const newUser = new User({
+      username,
+      email,
+      password_hash: hashedPassword,
+    });
+
+    await newUser.save();
+    res.status(201).json({ message: "Registration successful" });
+  } catch (error) {
+    res.status(500).json({ error: "Server error" });
+  }
+};
+
+export const loginUser = async (req, res) => {
+  try {
+    const { email, password } = req.body;
+    const user = await User.findOne({ email });
+    if (!user)
+      return res.status(400).json({ error: "Invalid email or password" });
+
+    const isMatch = await bcrypt.compare(password, user.password_hash);
+    if (!isMatch)
+      return res.status(400).json({ error: "Invalid email or password" });
+
+    res.status(201).json({
+      message: "Login successful",
+      user_id: user.id,
+      email: user.email,
+    });
+  } catch (error) {
+    res.status(500).json({ error: "Server error" });
+  }
+};
