+++ server/routes/profile.routes.js
@@ -2,13 +2,11 @@ import express from "express";
 import multer from "multer";
 import path from "path";
 import fs from "fs";
-import { fileURLToPath } from "url";
 
-const filePath = fileURLToPath(import.meta.url);
-const dirPath = path.dirname(filePath);
 const router = express.Router();
+const rootDir = process.cwd();
 
-const uploadDir = path.join(dirPath, "uploads");
+const uploadDir = path.join(rootDir, "uploads");
 if (!fs.existsSync(uploadDir)) {
   fs.mkdirSync(uploadDir, { recursive: true });
 }
@@ -28,15 +26,6 @@ router.post("/upload", upload.single("image"), (req, res) => {
 
   const fileUrl = `/uploads/${req.file.filename}`;
   res.json({ message: "Файл загружен", url: fileUrl });
-});
-
-router.get("/avatar/:filename", (req, res) => {
-  const filePath = path.join(uploadDir, req.params.filename);
-  if (fs.existsSync(filePath)) {
-    res.sendFile(filePath);
-  } else {
-    res.status(404).json({ message: "Файл не найден" });
-  }
 });
 
 export default router;
